sentiment_service/requirements.txt
grpcio==1.65.1
protobuf==5.26.1
